requirements.txt
click==8.0.3
cycler==0.11.0
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
fonttools==4.28.5
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mongoengine==0.23.1
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pymongo==3.12.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
WTForms==3.0.1
Flask-RESTful
gunicorn==19.6.0
Flask-Cors
pytest